payment/requirements.txt
aioredis==2.0.1
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
certifi==2021.10.8
charset-normalizer==2.0.12
cleo==1.0.0a4
click==8.0.4
crashtest==0.3.1
Deprecated==1.2.13
fastapi==0.75.0
h11==0.13.0
hiredis==2.0.0
idna==3.3
packaging==21.3
pptree==3.1
pydantic==1.9.0
pylev==1.4.0
pyparsing==3.0.7
python-dotenv==0.19.2
python-ulid==1.1.0
redis==4.2.0
redis-om==0.0.20
requests==2.27.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
types-redis==4.1.18
types-six==1.16.12
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
wrapt==1.14.0
